WondrousItems

+++ api/src/main/java/pl/kolendateam/dadcard/items/armor/repository/ArmorRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import pl.kolendateam.dadcard.items.armor.entity.Armors;
 
-public interface ArmorRepository extends JpaRepository<Armors, Short>{
+public interface ArmorRepository extends JpaRepository<Armors, Integer> {
 
 }
